+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## 0.0.6 - 2021-08-26?
 ### CLI
-* Re-organize some CLI commands:
-  * `create-usename` ==> `admin create-username`
+* Re-organized some CLI commands:
+  * `create-username` ==> `admin create-username`
   * `init` ==> `config init`
+  * Moved command registration to the `cli` package.
 
 ## 0.0.5 - 2021-08-23
 ### CLI